grpc_server/pyproject.toml
[tool.poetry]
name = "grpc-server"
version = "0.1.0"
description = ""
authors = ["VicenteMerino <vicente@trytoku.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
grpcio = "^1.60.0"
grpcio-tools = "^1.60.0"
sqlalchemy = "^2.0.23"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
grpclib = "^0.4.6"
protobuf = "^4.25.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
